+++ src/Lykke.Service.LiteCoin.API.Core/Transactions/ITransactionBuilderService.cs
@@ -5,7 +5,7 @@ namespace Lykke.Service.LiteCoin.API.Core.Transactions
 {
     public interface ITransactionBuilderService
     {
-        Task<Transaction> GetTransferTransaction(BitcoinAddress source, BitcoinAddress destination, decimal amount, bool sentDust = false);
+        Task<Transaction> GetTransferTransaction(BitcoinAddress source, BitcoinAddress destination, long amount, bool sentDust = false);
         Task<Transaction> GetSendMoneyToHotWalletTransaction(BitcoinAddress fromAddress, BitcoinAddress destination, string fromTxHash);
 
     }
